Реализовано приведение локальных объектов для передачи на сервер к виду понятному серверу.

+++ Toir-mobile/src/ru/toir/mobile/db/tables/EquipmentOperationResult.java
@@ -126,7 +126,7 @@ public class EquipmentOperationResult extends BaseTable {
 		this.type = type;
 	}
 
-	public long getType() {
+	public int getType() {
 		return type;
 	}
 }
